add image

+++ src/app/actions/character.action.ts
@@ -3,7 +3,7 @@
 import { createClient } from "@/supabase/server";
 import { revalidatePath } from "next/cache";
 
-export async function addCharacter(name: string, job: string) {
+export async function addCharacter(formData: FormData) {
   const supabase = createClient();
 
   const { data: auth, error: authError } = await supabase.auth.getUser();
@@ -12,11 +12,25 @@ export async function addCharacter(name: string, job: string) {
     return { success: false, message: "ไม่พบบัญชีผู้ใช้งาน" };
   }
 
-  const { error } = await supabase.from("characters").insert({
+  const image = formData.get("image") as File;
+
+  const dataObject = {
     user_id: auth.user.id,
-    name: name,
-    job: job,
-  });
+    name: formData.get("name") as string,
+    job: formData.get("job") as string,
+    image: image.name,
+    message_id: formData.get("messageId") as string,
+  };
+
+  const { error: imageError } = await supabase.storage
+    .from("character_image")
+    .upload(`private/${image.name}`, image);
+
+  if (imageError) {
+    return { success: false, message: "เพิ่มรูปภาพไม่สำเร็จ" };
+  }
+
+  const { error } = await supabase.from("characters").insert(dataObject);
 
   if (error) {
     return { success: false, message: "เพิ่มตัวละครไม่สำเร็จ" };
@@ -48,17 +62,24 @@ export async function getCharacters() {
     return { success: false, message: "ไม่พบตัวละคร", data: [] };
   }
 
-  revalidatePath("/characters");
   return { success: true, message: "", data: characters };
 }
 
-export async function deleteCharacter(charId: string) {
+export async function deleteCharacter(charId: string, oldImage: string) {
   const supabase = createClient();
 
   const { data: auth, error: authError } = await supabase.auth.getUser();
 
   if (!auth || authError) {
     return { success: false, message: "ไม่พบบัญชีผู้ใช้งาน" };
+  }
+
+  const { error: deleteError } = await supabase.storage
+    .from("character_image")
+    .remove([`private/${oldImage}`]);
+
+  if (deleteError) {
+    return { success: false, message: "แก้ไขรูปภาพไม่สำเร็จ" };
   }
 
   const { error } = await supabase.from("characters").delete().eq("id", charId);
@@ -69,4 +90,46 @@ export async function deleteCharacter(charId: string) {
 
   revalidatePath("/characters");
   return { success: true, message: "ลบตัวละครสำเร็จ" };
+}
+
+export async function updateCharImage(formData: FormData) {
+  const supabase = createClient();
+
+  const { data: auth, error: authError } = await supabase.auth.getUser();
+
+  if (!auth || authError) {
+    return { success: false, message: "ไม่พบบัญชีผู้ใช้งาน" };
+  }
+
+  const file = formData.get("image") as File;
+  const oldImage = formData.get("oldImage") as string;
+  const charId = formData.get("charId") as string;
+
+  const { error: deleteError } = await supabase.storage
+    .from("character_image")
+    .remove([`private/${oldImage}`]);
+
+  if (deleteError) {
+    return { success: false, message: "แก้ไขรูปภาพไม่สำเร็จ" };
+  }
+
+  const { error: imageError } = await supabase.storage
+    .from("character_image")
+    .upload(`/private/${file.name}`, file);
+
+  if (imageError) {
+    return { success: false, message: "แก้ไขรูปภาพไม่สำเร็จ" };
+  }
+
+  const { error } = await supabase
+    .from("characters")
+    .update({ image: file.name })
+    .eq("id", charId);
+
+  if (error) {
+    return { success: false, message: "แก้ไขชื่อไฟล์ไม่สำเร็จ" };
+  }
+
+  revalidatePath("/characters");
+  return { success: true, message: "แก้ไขรูปภาพสำเร็จ" };
 }
